# api/openapi.yaml
openapi: 3.1.0
info:
  title: Visto API
  version: 0.1.0
  description: Self-hosted movie and TV tracking API.
servers:
  - url: /api/v1
paths:
...
  /auth/google:
    get:
      summary: Start optional Google sign-in
      responses:
        '302': { description: 'Redirect to Google when OAuth is configured' }
        '404': { description: Google sign-in is not configured }
...
  /admin/backups:
    get:
      summary: Get instance backup settings and status as an administrator
      security: [{ sessionCookie: [] }, { bearerToken: [] }]
      responses:
        '200': { description: Backup settings without the secret access key, content: { application/json: { schema: { $ref: '#/components/schemas/BackupSettings' } } } }
        '403': { description: Administrator access required }
    put:
      summary: Save instance backup settings as an administrator
      security: [{ sessionCookie: [] }, { bearerToken: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/BackupSettingsRequest' } } } }
      responses:
        '200': { description: Saved backup settings without the secret access key, content: { application/json: { schema: { $ref: '#/components/schemas/BackupSettings' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { description: Administrator access required }
  /admin/backups/test:
    post:
      summary: Test saved S3 credentials as an administrator
      security: [{ sessionCookie: [] }, { bearerToken: [] }]
      responses:
        '200': { description: S3 connection works }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { description: Administrator access required }
  /admin/backups/run:
    post:
      summary: Start a manual backup using saved settings as an administrator
      security: [{ sessionCookie: [] }, { bearerToken: [] }]
      responses:
        '202': { description: Backup started }
        '403': { description: Administrator access required }
...
  /connected-apps:
    get:
      summary: List OAuth apps connected to the authenticated user's Visto account
      security: [{ sessionCookie: [] }, { bearerToken: [] }]
      responses:
        '200': { description: Connected OAuth apps, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ConnectedApp' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { description: OAuth is not configured }
    delete:
      summary: Revoke every OAuth app connected to the authenticated user's account
      security: [{ sessionCookie: [] }, { bearerToken: [] }]
      responses:
        '204': { description: All OAuth grants revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { description: OAuth is not configured }
  /connected-apps/{clientID}:
    delete:
      summary: Revoke one OAuth app connected to the authenticated user's account
      security: [{ sessionCookie: [] }, { bearerToken: [] }]
      parameters:
        - { name: clientID, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: OAuth grant revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { description: OAuth is not configured }
...
  /profile/plex-webhook:
    get:
      summary: Get this user's Plex webhook status and recent sync events
      security: [{ sessionCookie: [] }, { bearerToken: [] }]
      responses:
        '200': { description: Plex webhook status, content: { application/json: { schema: { $ref: '#/components/schemas/PlexWebhookStatus' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      summary: Create or rotate this user's Plex webhook URL; the secret is returned once
      security: [{ sessionCookie: [] }, { bearerToken: [] }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { account_id: { type: string, pattern: '^[0-9]*$', description: Numeric Plex account ID. An empty value skips all events until the webhook is rotated with an ID. } } } } } }
      responses:
        '201': { description: Created webhook URL, content: { application/json: { schema: { type: object, required: [webhook_url], properties: { webhook_url: { type: string, format: uri } } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { description: VISTO_PUBLIC_URL is not configured }
    delete:
      summary: Revoke this user's Plex webhook URL
      security: [{ sessionCookie: [] }, { bearerToken: [] }]
      responses:
        '204': { description: Webhook URL revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /webhooks/plex/{secret}:
    post:
      summary: Receive a Plex media.scrobble webhook
      parameters:
        - { name: secret, in: path, required: true, schema: { type: string, minLength: 40, maxLength: 80 } }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [payload]
              properties:
                payload: { type: string, description: Plex JSON event payload }
      responses:
        '204': { description: Event accepted or safely skipped }
        '400': { description: Invalid webhook form or missing payload }
        '404': { description: Unknown or revoked webhook secret }
        '413': { description: Webhook payload exceeds the 1 MiB limit }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: visto_session }
    bearerToken: { type: http, scheme: bearer, bearerFormat: Visto personal API token }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    ConnectedApp:
      type: object
      required: [client_id, client_name, scopes, expires_at]
      properties:
        client_id: { type: string }
        client_name: { type: string }
        scopes: { type: array, items: { type: string, enum: [read, write, offline_access] } }
        connected_at: { type: string, format: date-time }
        last_used_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
...
    BackupSettings:
      type: object
      properties:
        destination: { type: string, enum: [local, s3, both] }
        interval_seconds: { type: integer }
        bucket: { type: string }
        region: { type: string }
        endpoint: { type: string }
        access_key_id: { type: string }
        path_style: { type: boolean }
        prefix: { type: string }
        max_keep: { type: integer }
        has_secret: { type: boolean }
        encryption_available: { type: boolean }
        last_success_at: { type: string }
        last_error: { type: string }
    BackupSettingsRequest:
      allOf:
        - { $ref: '#/components/schemas/BackupSettings' }
        - type: object
          properties:
            secret_key: { type: string, writeOnly: true, description: Optional replacement secret access key }
...
    PlexWebhookEvent:
      type: object
      required: [id, status, occurred_at]
      properties:
        id: { type: integer, format: int64 }
        tmdb_id: { type: integer, format: int64 }
        status: { type: string, enum: [synced, skipped, failed] }
        title: { type: string }
        media_type: { type: string }
        message: { type: string }
        occurred_at: { type: string, format: date-time }
    PlexWebhookStatus:
      type: object
      required: [enabled, recent_events]
      properties:
        enabled: { type: boolean }
        account_id: { type: string, description: Numeric Plex account ID allowed to sync }
        created_at: { type: string, format: date-time }
        last_used_at: { type: string, format: date-time }
        last_synced_at: { type: string, format: date-time }
        recent_events: { type: array, maxItems: 20, items: { $ref: '#/components/schemas/PlexWebhookEvent' } }
...