# internal/presentation/http/server.go
package httpserver

import "net/http"
import "os"
import "path/filepath"
import "strings"
import "time"
import "github.com/afonsocosta/visto/internal/application/auth"
import "github.com/afonsocosta/visto/internal/infrastructure/backup"
import exportapp "github.com/afonsocosta/visto/internal/application/export"
import "github.com/afonsocosta/visto/internal/application/feed"
import "github.com/afonsocosta/visto/internal/application/library"
import "github.com/afonsocosta/visto/internal/application/oauth"
import "github.com/afonsocosta/visto/internal/application/plexsync"
import "github.com/afonsocosta/visto/internal/application/profile"
import "github.com/afonsocosta/visto/internal/application/tracking"
import "github.com/afonsocosta/visto/internal/application/watch"
import "github.com/afonsocosta/visto/internal/domain"
import "github.com/afonsocosta/visto/internal/presentation/security"

type Server struct {
	handler     http.Handler
	mux         *http.ServeMux
	authService *auth.Service
	proxies     *security.ProxyResolver
}

func New(authService *auth.Service, metadataProvider domain.MetadataProvider, webDir string, libraryService *library.Service, trackingService *tracking.Service, profileService *profile.Service, feedService *feed.Service, exportService *exportapp.Service, watchService *watch.Service, webPush ...*WebPushConfig) *Server {
	mux := http.NewServeMux()
	proxies := &security.ProxyResolver{}
	loginLimiter := newLoginLimiter()
	registerAPIRoutes(mux, authService, metadataProvider, libraryService, trackingService, profileService, feedService, exportService, watchService, loginLimiter, proxies)
	var pushConfig *WebPushConfig
	if len(webPush) > 0 {
		pushConfig = webPush[0]
	}
	pushRoutes(mux, authService, pushConfig)
	if webDir != "" {
		if _, err := os.Stat(webDir); err == nil {
			mux.Handle("GET /", singlePageApp(webDir))
		}
	}
	return &Server{handler: limitAPIRequestBody(csrfProtection(mux, proxies)), mux: mux, authService: authService, proxies: proxies}
}

func limitAPIRequestBody(next http.Handler) http.Handler {
	const maxBody = 1 << 20
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if strings.HasPrefix(r.URL.Path, "/api/v1/") && !strings.HasPrefix(r.URL.Path, "/api/v1/webhooks/plex/") && r.Method != http.MethodGet && r.Method != http.MethodHead {
			if r.ContentLength > maxBody {
				http.Error(w, "request body too large", http.StatusRequestEntityTooLarge)
				return
			}
			r.Body = http.MaxBytesReader(w, r.Body, maxBody)
		}
		next.ServeHTTP(w, r)
	})
}

func (server *Server) WithTrustedProxies(proxies security.ProxyResolver) *Server {
	if server.proxies != nil {
		*server.proxies = proxies
	}
	return server
}

// WithOAuth adds account-management endpoints for OAuth clients. It is kept
// optional so Visto can run without an OAuth configuration.
func (server *Server) WithOAuth(service *oauth.Service) *Server {
	server.mux.HandleFunc("GET /api/v1/connected-apps", listConnectedApps(server.authService, service))
	server.mux.HandleFunc("DELETE /api/v1/connected-apps", revokeAllConnectedApps(server.authService, service))
	server.mux.HandleFunc("DELETE /api/v1/connected-apps/{clientID}", revokeConnectedApp(server.authService, service))
	return server
}

// WithGoogleOAuth adds optional Google sign-in routes when all credentials are set.
func (server *Server) WithGoogleOAuth(config GoogleOAuthConfig) *Server {
	if config.Enabled() {
		server.mux.HandleFunc("GET /api/v1/auth/google", googleLogin(config, server.authService, server.proxies))
		server.mux.HandleFunc("GET /api/v1/auth/google/callback", googleCallback(config, server.authService, server.proxies))
	}
	return server
}

// WithPlexSync adds per-user Plex webhook management and the public callback.
func (server *Server) WithPlexSync(service *plexsync.Service) *Server {
	if service == nil {
		return server
	}
	server.mux.HandleFunc("GET /api/v1/profile/plex-webhook", plexWebhookStatus(server.authService, service))
	server.mux.HandleFunc("POST /api/v1/profile/plex-webhook", issuePlexWebhook(server.authService, service))
	server.mux.HandleFunc("DELETE /api/v1/profile/plex-webhook", revokePlexWebhook(server.authService, service))
	server.mux.HandleFunc("POST /api/v1/webhooks/plex/{secret}", plexWebhook(service))
	return server
}

func (server *Server) WithBackups(service *backup.Service) *Server {
	server.mux.HandleFunc("GET /api/v1/admin/backups", adminBackup(server.authService, service, "get"))
	server.mux.HandleFunc("PUT /api/v1/admin/backups", adminBackup(server.authService, service, "save"))
	server.mux.HandleFunc("POST /api/v1/admin/backups/test", adminBackup(server.authService, service, "test"))
	server.mux.HandleFunc("POST /api/v1/admin/backups/run", adminBackup(server.authService, service, "run"))
	return server
}

func singlePageApp(webDir string) http.Handler {
	files := http.FileServer(http.Dir(webDir))
	index := filepath.Join(webDir, "index.html")
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		requested := filepath.Join(webDir, filepath.Clean("/"+r.URL.Path))
		if info, err := os.Stat(requested); err == nil && !info.IsDir() {
			files.ServeHTTP(w, r)
			return
		}
		http.ServeFile(w, r, index)
	})
}

func listConnectedApps(authService *auth.Service, service *oauth.Service) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		user, ok := authenticatedUser(w, r, authService)
		if !ok {
			return
		}
		if service == nil {
			writeError(w, http.StatusServiceUnavailable, "connected apps are not configured")
			return
		}
		connections, err := service.Connections(r.Context(), user.ID)
		if err != nil {
			writeError(w, http.StatusInternalServerError, "could not list connected apps")
			return
		}
		writeJSON(w, http.StatusOK, connections)
	}
}

func revokeConnectedApp(authService *auth.Service, service *oauth.Service) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		user, ok := authenticatedUser(w, r, authService)
		if !ok {
			return
		}
		if service == nil {
			writeError(w, http.StatusServiceUnavailable, "connected apps are not configured")
			return
		}
		if err := service.RevokeConnection(r.Context(), user.ID, r.PathValue("clientID")); err != nil {
			writeError(w, http.StatusBadRequest, "could not revoke connected app")
			return
		}
		w.WriteHeader(http.StatusNoContent)
	}
}

func revokeAllConnectedApps(authService *auth.Service, service *oauth.Service) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		user, ok := authenticatedUser(w, r, authService)
		if !ok {
			return
		}
		if service == nil {
			writeError(w, http.StatusServiceUnavailable, "connected apps are not configured")
			return
		}
		if err := service.RevokeAllConnections(r.Context(), user.ID); err != nil {
			writeError(w, http.StatusInternalServerError, "could not revoke connected apps")
			return
		}
		w.WriteHeader(http.StatusNoContent)
	}
}

func (server *Server) Handler() http.Handler {
	return server.handler
}

func health(response http.ResponseWriter, _ *http.Request) {
	writeJSON(response, http.StatusOK, map[string]string{
		"status": "ok",
		"time":   time.Now().UTC().Format(time.RFC3339),
	})
}